requirements.txt
pyserial
python-escpos
requests
pyusb
jinja2
Pillow
ollama
pandas
openpyxl
cairosvg
drawsvg